DataProcess를 새로운 파일로 이동
