+++ src/main/java/dev/hoto/iotmanage/controller/DeviceController.java
@@ -1,5 +1,6 @@
 package dev.hoto.iotmanage.controller;
 
+import dev.hoto.iotmanage.DataProcess;
 import dev.hoto.iotmanage.IoTDevice;
 import dev.hoto.iotmanage.service.deviceServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -111,47 +112,6 @@ public class DeviceController {
 
         deviceService.DeleteDeviceById(moduleId); // 해당 디바이스 제거
         return "redirect:/show"; // show 페이지로 가기
-    }
-}
-
-class DataProcess {
-    public static void changeModuleTypeToKorean(List<IoTDevice> target) {
-        String[] korTypes = {"전등", "창문", "블라인드", "가스"};
-        int index = -1;
-
-        for (IoTDevice device :
-                target) {
-            changeModuleTypeToKorean(device);
-        }
-    }
-
-    public static void changeModuleTypeToKorean(IoTDevice device) {
-        String[] korTypes = {"전등", "창문", "블라인드", "가스"};
-        int index = -1;
-
-        switch (device.getType()) {
-            case "light":
-                index = 0;
-                break;
-
-            case "window":
-                index = 1;
-                break;
-
-            case "blind":
-                index = 2;
-                break;
-
-            case "gas":
-                index = 3;
-                break;
-
-            default:
-                System.out.println("잘못된 모듈 종류입니다.");
-                System.exit(0);
-        }
-
-        device.setType(korTypes[index]);
     }
 }
